fix(admin_func): use prepared statements to prevent SQL injection
- Replace all direct SQL queries with prepared statements in admin_func.php
- Use parameterized queries for DELETE operations on users table
- Use parameterized queries for UPDATE operations on deletion_requests table
- Use parameterized queries for SELECT operations with user_id and item_id filters
- Bind integer and decimal parameters correctly to prevent type coercion attacks
- Apply the same security fixes to market history reversion logic
- Apply the same security fixes to trade history reversion logic
- Update gearup.sql, history.php, home.php, index.php, and market.php for consistency
- Eliminates vulnerability to SQL injection attacks through unsanitized user input

+++ gearup.sql
@@ -312,7 +312,7 @@ CREATE TABLE `users` (
   `picture` varchar(255) NOT NULL,
   `email` varchar(255) NOT NULL,
   `password` varchar(255) NOT NULL,
-  `credits` int(11) NOT NULL DEFAULT 10000,
+  `credits` decimal(10,2) NOT NULL DEFAULT 10000.00,
   `is_admin` tinyint(1) DEFAULT 0,
   PRIMARY KEY (`id`),
   UNIQUE KEY `name` (`name`),
